Add deployComponent
